Small bug fixes and added automatic pull for terms and outlets

+++ src/worker/api_wrapper.py
@@ -2,6 +2,7 @@ import requests
 import json
 from datetime import datetime
 import gzip
+import os
 import io
 import zipfile
 
@@ -17,7 +18,19 @@ def push_new(path='resources/examples/test_crawler_py.json'):
 
 def fetch_results(
         path=f'resources/raw_data/{datetime.today().date().isoformat()}.json'):
-    raw_data = requests.get(DOMAIN + 'queues-processed')
+    if os.path.exists(path):
+        raise Exception('file already exists')
+    raw_data = requests.get(DOMAIN + 'queues-unreviewed')
+    raw_dict = raw_data.json()
+    with open(path, 'w') as f:
+        json.dump(raw_dict, f, indent='  ')
+
+
+def fetch_reviewed(
+        path=f'resources/raw_data/reviewed_{datetime.today().date().isoformat()}.json'):
+    if os.path.exists(path):
+        raise Exception('file already exists')
+    raw_data = requests.get(DOMAIN + 'queues-reviewed')
     raw_dict = raw_data.json()
     with open(path, 'w') as f:
         json.dump(raw_dict, f, indent='  ')
@@ -42,3 +55,21 @@ def fetch_html(url):
     raw_html = as_zipfile.read(name)
 
     return raw_html
+
+
+def set_reviewed(queue_id):
+    r = requests.put(DOMAIN + f'queues/{queue_id}', data={'reviewed': 1})
+    if r.status_code == 200:
+        return 'success'
+    else:
+        raise ConnectionError(f'status code {r.status_code}')
+
+
+def get_outlets():
+    r = requests.get(DOMAIN + 'outlets')
+    return r.json()['data']
+
+
+def get_terms():
+    r = requests.get(DOMAIN + 'terms')
+    return r.json()['data']
